requirements.txt
google-api-python-client==2.114.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
flask==3.0.0
pygments==2.17.2
click==8.1.7
py7zr==0.21.0
openai>=1.0.0
